Completed create, delete, and read actions for a single club.
Added create, delete functions to api/clubs.js, and added corresponding create/destroy functions in Clubs.js and Club.js, respectively. Broke out the rendering of a single club to it's own component, which is now shared when viewing a single club or the fully indexed list. Created a ClubForm component that is used to create a new club from the /clubs view, which will also be used for updating clubs.

Follow-up will be to figure out how to submit a club stiffness rating from the drop-down <option> element.